--- src/const.d.ts
@@ -1,1 +1,0 @@
-// declare const TIMEOUT_FACTOR: number;
