+++ src/index.ts
@@ -1,6 +1,6 @@
-export {Waterfall} from "./Waterfall";
-export {IWaterfallOptions} from "./IWaterfallOptions";
-export {exponentialTruncatedBackoff} from "./exponentialTruncatedBackoff";
-export {Pipe} from "./Pipe";
-export {Dam} from "./Dam";
-export {Tank} from "./Tank";
+export {Waterfall} from "./waterfall/Waterfall";
+export {IWaterfallOptions} from "./waterfall/IWaterfallOptions";
+export {exponentialTruncatedBackoff} from "./waterfall/exponentialTruncatedBackoff";
+export {Pipe} from "./pipe/Pipe";
+export {Dam} from "./dam/Dam";
+export {Tank} from "./tank/Tank";
