+++ src/polyfill/CloseEvent.ts
