+++ src/polyfill/isNode.ts
